Why would it not be like this.
Doing this because, to me, it looks cleaner.

It appears that deep in the code there is a default option for settings.options.bindEvents.
But it isn't clear that it is there because defaults.options.bindEvents is an empty array.
Why would there be the conditional to provide the default? I can see the only difference is handling the case of  if settings.options.bindEvents is [].

Someone can explain to me why it was the way it was.